+++ fastats/maths/correlation/__init__.py
@@ -1,0 +1,9 @@
+
+from fastats.maths.correlation.pearson import pearson
+from fastats.maths.correlation.spearman import spearman
+
+
+__all__ = [
+    'pearson',
+    'spearman',
+]