+++ fastats/maths/correlation/spearman.py
@@ -1,0 +1,42 @@
+
+import numpy as np
+
+
+def spearman(x, y):
+    """
+    Calculates the Spearman rank correlation
+    coefficient for the inputs `x` and `y`.
+
+    If there are any Nan values in the data,
+    they will be ignored in the rank for that
+    variable, and therefore may skew the results.
+    See the test_spearman unit-tests for an
+    example.
+
+    Example
+    -------
+
+    >>> import numpy as np
+    >>> x = np.array([1, 2, 3, 4])
+    >>> y = np.array([2, 3, 4, 3])
+    >>> spearman(x, y)  # doctest: +ELLIPSIS
+    0.800000000...
+
+    :param x: A `numpy.array` of floats or ints
+    :param y: A `numpy.array` of floats or ints
+    :return: A float representing the correlation
+    """
+    assert x.shape == y.shape
+
+    n = len(x)
+    rank_x = np.argsort(x)
+    rank_y = np.argsort(y)
+
+    d = rank_x - rank_y
+    d2 = d ** 2
+    return 1 - (6 * np.sum(d2)) / (n**3 - n)
+
+
+if __name__ == '__main__':
+    import pytest
+    pytest.main([__file__])