+++ releases.md
@@ -4,6 +4,7 @@
 #### New features
 
 - erf : error function using Abramowitz and Stegun method (maximum error: 1.5e-7) 
+- correlation : Spearman and Pearson correlation coefficient functions
 
 #### Bug fixes